requirements.txt
supabase==2.27.0
python-dotenv==1.2.1
